+++ R/docx_exporter_functions.R
@@ -3,6 +3,23 @@ dps_markup_df_docx <- tibble::tibble(
   replace_by = c("flextable::as_sup", "flextable::as_sub")
 )
 
+validate_tabletree <- function(tt, validate, tlgtype) {
+  if (validate && tlgtype == "Table" && methods::is(tt, "VTableTree")) {
+    if (!rtables::validate_table_struct(tt)) {
+      message(
+        "Invalid table structure detected. This may cause issues in the output. ",
+        "The validation process failed, proceed with caution."
+      )
+    }
+  } else if (!validate && tlgtype == "Table" && methods::is(tt, "VTableTree")) {
+    if (rtables::validate_table_struct(tt)) {
+      message(
+        "Table structure validation succeeded. You should not need to set validate=FALSE."
+      )
+    }
+  }
+}
+
 get_template_file <- function() {
   template_file <- "template_file.docx"
   template_file <- system.file(template_file, package = "junco")
@@ -567,7 +584,7 @@ interpret_all_cell_content <- function(flx, markup_df_docx = dps_markup_df_docx)
 #' (optional) Default = NULL.
 #'
 #' @inheritParams export_as_docx_j
-#' @inherit export_TLG_as_docx note
+#' @inheritSection export_TLG_as_docx Note
 #'
 #' @returns A function that applies the given theme to a flextable.
 #' @export
@@ -704,11 +721,11 @@ theme_docx_default_j <- function(
 #' the indentation if we have vertical pagination. No need to be specified by the end user.\cr
 #' (optional) Default = FALSE.
 #'
-#' @inheritParams export_as_docx_j
-#' @inherit export_TLG_as_docx note
+#' @inheritParams export_TLG_as_docx
+#' @inheritSection export_TLG_as_docx Note
 #'
 #' @note
-#' IMPORTANT NOTE: the following features are not implemented in `flextable`,
+#' The following features are not implemented in `flextable`,
 #' and as a result they will only be visible when exporting to docx using
 #' [junco::export_TLG_as_docx()]:
 #' - watermark
@@ -745,6 +762,7 @@ tt_to_flextable_j <- function(
     round_type = formatters::obj_round_type(tt),
     alignments = list(),
     border_mat = make_header_bordmat(obj = tt),
+    validate = TRUE,
     ...) {
   if (inherits(tt, "list")) {
     stop("Please use paginate = TRUE or mapply() to create multiple outputs. export_as_docx accepts lists.")
@@ -784,6 +802,8 @@ tt_to_flextable_j <- function(
   for (alignment in alignments) {
     stopifnot("Each item of `alignments` must be a list" = is.list(alignment))
   }
+
+  validate_tabletree(tt, validate, tlgtype)
 
   if (tlgtype == "Listing" && nrow(tt) == 0) {
     dat <- as.list(c("No data to report", rep("", ncol(tt) - 1)))
@@ -953,6 +973,7 @@ tt_to_flextable_j <- function(
           alignments = alignments,
           border_mat = pag_bord_mats[[i]],
           export_csv = FALSE, # this is because we already exported the csv a few lines above,
+          validate = validate,
           ...
         )
 
@@ -1321,7 +1342,7 @@ tt_to_flextable_j <- function(
       tt,
       fontspec = fontspec,
       landscape = orientation == "landscape",
-      colwidths = NULL,
+      colwidths = colwidths_2,
       col_gap = col_gap,
       pg_width = pg_width_by_orient(orientation == "landscape"),
       pg_height = NULL,
@@ -1393,62 +1414,9 @@ tt_to_flextable_j <- function(
 #' See notes to understand why this is experimental.
 #'
 #' @param tt (`TableTree` or `listing_df`)\cr the object to export.
-#' @param tblid (`character`)\cr Output ID to be displayed in the title and last line of footer.
-#' @param output_dir (`character`)\cr a directory path to save the docx.
-#' @param theme (function factory)\cr The theme to apply to the flextable
-#' (default = [theme_docx_default_j()]).\cr
-#' See [theme_docx_default_j()] or [rtables.officer::theme_docx_default()]
-#' for more details.
-#' @param add_page_break (`logical`)\cr (optional) Default = FALSE.
-#' @param titles_as_header (`logical`)\cr (optional) Default = TRUE.
-#' @param integrate_footers (`logical`)\cr (optional) Default = TRUE.
-#' @param section_properties (`prop_section`)\cr (optional) A "prop_section" object
-#' containing information about page size, orientation, margins, etc.
-#' See [officer::prop_section()] for more details.
-#' No need to be specified by end user.
-#' @param doc_metadata (list of `string`)\cr Any value that can be used as metadata
-#' by [officer::set_doc_properties()]. Important text values are title, subject,
-#' creator, and description, while created is a date object.\cr
-#' (optional) Default = NULL.
-#' @param orientation (`character`)\cr (optional) Default = "portrait".
-#' One of: "portrait", "landscape".
-#' @param paginate (`logical`)\cr (optional) Default = TRUE for TableTree and FALSE otherwise.
-#' @param nosplitin (`list`)\cr list(row=, col=). Path elements whose children should not be paginated within
-#' if it can be avoided. e.g., list(col="TRT01A") means don't split within treatment arms unless
-#' all the associated columns don't fit on a single page.
-#' @param string_map (`tibble`)\cr (optional) Default = default_str_map.
-#' @param markup_df_docx (`tibble`)\cr (optional) Default = dps_markup_df_docx.
-#' @param combined_docx (`logical`)\cr Whether to also export an "allparts" docx version.\cr
-#' (optional) Default = FALSE.
-#' @param tlgtype (`character`)\cr (optional). No need to be specified by end user.
-#' @param col_gap (`numeric`)\cr (optional). Default = 3 (Tables) or 0.5 (Listings).
-#' @param pagenum (`logical`)\cr (optional). Whether to display page numbers. Only applicable
-#' to listings (i.e. for tables and figures this argument is ignored).
-#' @param round_type (`"iec"` or `"sas"`)\cr the type of rounding to perform. iec,
-#' the default, performs rounding compliant with IEC 60559, while
-#' sas performs nearest-value rounding consistent with rounding within SAS.
-#' See `[formatters::format_value()]` for more details.
-#' @param alignments (`list`)\cr (optional) List of named lists. Vectorized.
-#' (Default = `list()`) Used to specify individual column or cell alignments.
-#' Each named list contains `row`, `col`, and `value`.
-#' @param border (`fp_border`)\cr Border to use (default =
-#' \code{flextable::fp_border_default(width = 0.75, color = "black")}).
-#' @param border_mat (`matrix`)\cr A `m x k` matrix where m is the number of columns of `tt`
-#' and k is the number of lines the header takes up. See [tidytlg::add_bottom_borders]
-#' for what the matrix should contain. Users should only specify this when the
-#' default behavior does not meet their needs.
-#' @param export_csv (`logical(1)`)\cr Whether to export the object as a csv representation.
-#' Default = FALSE.
-#' @param output_csv_directory (`character(1)`)\cr the directory to export the csv.
-#' Default = NULL. Only used if export_csv = TRUE.
-#' If NULL or attempting to export in a non-existent directory, the csv will be exported
-#' in the same directory as the .docx file.
-#' @param markup_df (`data.frame`)\cr Data frame containing markup information.
-#' Only used if export_csv = TRUE.
-#' @param ... other parameters.
 #'
 #' @inheritParams export_TLG_as_docx
-#' @inherit export_TLG_as_docx note
+#' @inheritSection export_TLG_as_docx Note
 export_as_docx_j <- function(
   tt,
   tblid,
@@ -1487,6 +1455,7 @@ export_as_docx_j <- function(
   export_csv = FALSE,
   output_csv_directory = NULL,
   markup_df = dps_markup_df,
+  validate = TRUE,
   ...
 ) {
 
@@ -1507,10 +1476,10 @@ export_as_docx_j <- function(
   checkmate::assert_choice(orientation, choices = c("portrait", "landscape"))
   checkmate::assert_flag(paginate)
   checkmate::assert_list(nosplitin)
-  checkmate::assert_multi_class(x = string_map,
-                                classes = c("tbl_df", "data.frame"))
-  checkmate::assert_multi_class(x = markup_df_docx,
-                                classes = c("tbl_df", "data.frame"))
+  checkmate::assert_tibble(string_map)
+  checkmate::assert_true(all(colnames(string_map) == c("pattern", "value")))
+  checkmate::assert_tibble(markup_df_docx)
+  checkmate::assert_true(all(colnames(markup_df_docx) == c("keyword", "replace_by")))
   checkmate::assert_flag(combined_docx)
   checkmate::assert_character(tlgtype)
   checkmate::assert_numeric(col_gap)
@@ -1519,6 +1488,8 @@ export_as_docx_j <- function(
   checkmate::assert_list(alignments)
   checkmate::assert_class(border, "fp_border")
   checkmate::assert_matrix(border_mat)
+  checkmate::assert_tibble(markup_df)
+  checkmate::assert_true(all(colnames(markup_df) == c("keyword", "rtfstart", "rtfend")))
 
   # Validate `alignments` here because of its complicated data structure
   stopifnot("`alignments` must be a list" = is.list(alignments))
@@ -1557,6 +1528,7 @@ export_as_docx_j <- function(
       output_csv_directory = output_csv_directory,
       markup_df = markup_df,
       output_dir = output_dir, # this argument is needed to guess where to save the csv
+      validate = validate,
       ...
     )
   }
@@ -1590,6 +1562,7 @@ export_as_docx_j <- function(
             output_csv_directory = output_csv_directory,
             markup_df = markup_df,
             output_dir = output_dir, # this argument is needed to guess where to save the csv
+            validate = validate,
             ...
           ),
         SIMPLIFY = FALSE
@@ -1647,6 +1620,7 @@ export_as_docx_j <- function(
         export_csv = export_csv,
         output_csv_directory = output_csv_directory,
         markup_df = markup_df,
+        validate = validate,
         ...
       )
     } else {
@@ -1690,6 +1664,7 @@ export_as_docx_j <- function(
         export_csv = export_csv,
         output_csv_directory = output_csv_directory,
         markup_df = markup_df,
+        validate = validate,
         ...
       )
     }
@@ -1823,7 +1798,7 @@ export_as_docx_j <- function(
 #' (optional) Default = NULL.
 #'
 #' @inheritParams export_TLG_as_docx
-#' @inherit export_TLG_as_docx note
+#' @inheritSection export_TLG_as_docx Note
 export_graph_as_docx <- function(g = NULL,
                                  plotnames = NULL,
                                  tblid,
@@ -2063,6 +2038,17 @@ export_graph_as_docx <- function(g = NULL,
 #' the input Table/Listing and k is the number of lines the header takes up.\cr
 #' See [tidytlg::add_bottom_borders] for what the matrix should contain.
 #' Users should only specify this when the default behavior does not meet their needs.
+#' @param export_csv (`logical(1)`)\cr Whether to export the object as a csv representation.
+#' Default = FALSE.
+#' @param output_csv_directory (`character(1)`)\cr the directory to export the csv.
+#' Default = NULL. Only used if export_csv = TRUE.
+#' If NULL or attempting to export in a non-existent directory, the csv will be exported
+#' in the same directory as the .docx file.
+#' @param markup_df (`data.frame`)\cr Data frame containing markup information.
+#' Only used if export_csv = TRUE.
+#' @param validate (`logical(1)`)\cr Whether to validate the table structure using
+#'  `rtables::validate_table_struct()`. Defaults to `TRUE`. If `FALSE`, a message
+#'  will be displayed when validation fails.
 #' @param watermark (`character`)\cr the watermark (text) to display in the
 #' output docx file.\cr
 #' If NULL, no watermark will be displayed.
@@ -2152,7 +2138,7 @@ export_graph_as_docx <- function(g = NULL,
 #'
 #' @export
 #'
-#' @note
+#' @section Note:
 #' This function has been tested for common use cases but may not work or have
 #' unexpected or undesired behavior in corner cases. As such it is not considered
 #' fully production ready and is being made available for further testing
@@ -2196,6 +2182,10 @@ export_TLG_as_docx <- function(
   alignments = list(),
   border = flextable::fp_border_default(width = 0.75, color = "black"),
   border_mat = NULL,
+  export_csv = FALSE,
+  output_csv_directory = NULL,
+  markup_df = dps_markup_df,
+  validate = TRUE,
   watermark = NULL,
   plotnames = NULL,
   title = NULL,
@@ -2232,10 +2222,10 @@ export_TLG_as_docx <- function(
   checkmate::assert_choice(orientation, choices = c("portrait", "landscape"))
   checkmate::assert_flag(paginate)
   checkmate::assert_list(nosplitin)
-  checkmate::assert_multi_class(x = string_map,
-                                classes = c("tbl_df", "data.frame"))
-  checkmate::assert_multi_class(x = markup_df_docx,
-                                classes = c("tbl_df", "data.frame"))
+  checkmate::assert_tibble(string_map)
+  checkmate::assert_true(all(colnames(string_map) == c("pattern", "value")))
+  checkmate::assert_tibble(markup_df_docx)
+  checkmate::assert_true(all(colnames(markup_df_docx) == c("keyword", "replace_by")))
   checkmate::assert_flag(combined_docx)
   checkmate::assert_character(tlgtype)
   checkmate::assert_numeric(col_gap)
@@ -2244,6 +2234,11 @@ export_TLG_as_docx <- function(
   checkmate::assert_list(alignments)
   checkmate::assert_class(border, "fp_border")
   checkmate::assert_matrix(border_mat)
+  checkmate::assert_flag(export_csv)
+  checkmate::assert_character(output_csv_directory, null.ok = TRUE, len = 1)
+  checkmate::assert_tibble(markup_df)
+  checkmate::assert_true(all(colnames(markup_df) == c("keyword", "rtfstart", "rtfend")))
+  checkmate::assert_flag(validate)
   checkmate::assert_character(watermark, len = 1, null.ok = TRUE)
   checkmate::assert_character(plotnames, null.ok = TRUE)
   checkmate::assert_character(title, null.ok = TRUE)
@@ -2265,7 +2260,11 @@ export_TLG_as_docx <- function(
       combined_docx = combined_docx, tlgtype = tlgtype,
       col_gap = col_gap, pagenum = pagenum, round_type = round_type,
       alignments = alignments, border = border,
-      border_mat = border_mat, watermark = watermark, ...
+      border_mat = border_mat, watermark = watermark,
+      export_csv = export_csv, output_csv_directory = output_csv_directory,
+      markup_df = markup_df,
+      validate = validate,
+      ...
     )
   } else {
     export_graph_as_docx(g = obj, plotnames = plotnames,